Move mail-service into server/src/domain package
