+++ server/test/integration/routes/feedbacks.spec.js
@@ -1,7 +1,7 @@
 const jwt = require('jsonwebtoken');
 const { request, expect, sinon } = require('../../test-helper');
 const app = require('../../../app');
-const mailService = require('../../../src/infrastructure/mail-service');
+const mailService = require('../../../src/domain/mail-service');
 const slackService = require('../../../src/domain/slack-service');
 
 describe('Integration | Routes | feedbacks route', () => {
